refactor: inline wand popup logic
