+++ src/image/stage_toolbar/index.js
@@ -12,6 +12,7 @@ import redo from './redo.svg';
 import path from './path.svg';
 import direction from './direction.svg';
 import settings from './settings.svg';
+import wand from './wand.svg';
 
 export default {
   stroke,
@@ -28,4 +29,5 @@ export default {
   undo,
   redo,
   settings,
+  wand,
 };
